// client-app/src/app/models/user.ts
export interface User {
    username: string;
    displayName: string;
    token: string;
    image?: string;
}

export interface UserFormValues {
    email: string;
    password: string;
    displayName?: string;
    username?: string;
}

export interface userFormError {
    email: string;
    password: string;
    displayName?: string;
    username?: string;
}